fix(@ngtools/webpack): only load test helpers on tests
Fix #16858
