+++ packages/ngtools/webpack/src/transformers/multiple_transformers_spec.ts
@@ -6,11 +6,11 @@
  * found in the LICENSE file at https://angular.io/license
  */
 import { tags } from '@angular-devkit/core';  // tslint:disable-line:no-implicit-dependencies
-import { createTypescriptContext, transformTypescript } from './ast_helpers';
 import { exportLazyModuleMap } from './export_lazy_module_map';
 import { exportNgFactory } from './export_ngfactory';
 import { removeDecorators } from './remove_decorators';
 import { replaceBootstrap } from './replace_bootstrap';
+import { createTypescriptContext, transformTypescript } from './spec_helpers';
 
 
 describe('@ngtools/webpack transformers', () => {
